update int to double
waiting for edit draw()

+++ src/logic/Hitbox.java
@@ -9,8 +9,8 @@ public abstract class Hitbox {
 	protected Position C;
 	//down-right
 	protected Position D;
-	protected int xSpeed;
-	protected int ySpeed;
+	protected double xSpeed;
+	protected double ySpeed;
 	
 	public Position getA() {
 		return A;
@@ -24,11 +24,11 @@ public abstract class Hitbox {
 		return B;
 	}
 
-	public int getySpeed() {
+	public double getySpeed() {
 		return ySpeed;
 	}
 
-	public void setySpeed(int ySpeed) {
+	public void setySpeed(double ySpeed) {
 		this.ySpeed = ySpeed;
 	}
 
@@ -52,16 +52,16 @@ public abstract class Hitbox {
 		D = d;
 	}
 
-	public int getxSpeed() {
+	public double getxSpeed() {
 		return xSpeed;
 	}
 
-	public void setxSpeed(int xSpeed) {
+	public void setxSpeed(double xSpeed) {
 		this.xSpeed = xSpeed;
 	}
 
 	//Obstacle has only X-velocity
-	public Hitbox(Position position, int speed) {
+	public Hitbox(Position position, double speed) {
 		this.xSpeed = speed;
 		
 	}
@@ -76,7 +76,7 @@ public abstract class Hitbox {
 	}
 	
 	//Monster has XY-velocity
-	public Hitbox(Position a,int width,int height,int xSpeed, int ySpeed) {
+	public Hitbox(Position a,int width,int height,double xSpeed, double ySpeed) {
 		super();
 		Position A = new Position(a.getX(), a.getY());
 		Position B = new Position(a.getX(), a.getY()+height);
@@ -97,10 +97,10 @@ public abstract class Hitbox {
 		return false;
 		}
 	
-	public void update(int xSpeed, int ySpeed, double time) {
+	public void update(double xSpeed, double ySpeed, long time) {
 		for(Position i  : new Position[] {this.A, this.B, this.C, this.D}){
-			i.setX(i.getX() + xSpeed * (int)time);
-			i.setY(i.getY() + ySpeed * (int)time);
+			i.setX(i.getX() + xSpeed * time);
+			i.setY(i.getY() + ySpeed * time);
 		}
 
 	}
